requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
alembic==1.13.3
asyncpg==0.30.0
redis==5.2.0
celery[redis]==5.4.0
celery[beat]==5.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
Pillow==11.0.0
aiogram==3.13.1
aiofiles==24.1.0
fastapi-pagination==0.12.29
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2